+++ Lora/Events/RecievedData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Fraunhofer.Fit.Iot.Lora.Events {
+  public class RecievedData {
+    public Byte[] Data {
+      get; set;
+    }
+  }
+}