+++ Lora/lib/Dragino/Dragino.Config.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Threading;
+
 using BlubbFish.Utils;
+
 using Unosquare.RaspberryIO;
 using Unosquare.WiringPi;