+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Changelog
-## 1.9.0
+## 2.0.0 - The total rewrite
+
 ### New Features
+* Dragino board now support Sending data while listen on receiving data
+* Now netcore is supported
+
 ### Bugfixes
+
 ### Changes
+* Rewrite LoraController and the rest so that it did not parse any data
+* Rewrite Dragino driver complete
+* Remove all parsing data stuff, that must now be done in the programm that call the lib
 * Split Ic800A into different files
 * Fix Thread abort in netcore
+* Using new WiringPi lib
 
 ## 1.8.4
 ### New Features